requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
black==24.10.0
celery==5.4.0
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.7
django-filter==24.3
djangorestframework==3.15.2
docx2pdf==0.1.8
et-xmlfile==1.1.0
flake8==7.1.1
kombu==5.4.0
lxml==5.3.0
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.0.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.2
pathspec==0.12.1
pdfkit==1.0.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
psycopg2==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
pypandoc==1.14
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
pytz==2024.1
pywin32==308
reportlab==4.2.5
six==1.16.0
sqlparse==0.5.1
tomli==2.1.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
wkhtmltopdf==0.2
XlsxWriter==3.2.0
